Determine khworkload once; workload cannot be retrieved from spec
