+++ pkg/checks/external/main.go
@@ -116,6 +116,7 @@ type Checker struct {
 	wg                       sync.WaitGroup     // used to track background workers and processes
 	hostname                 string             // hostname cache
 	checkPodName             string             // the current unique checker pod name
+	KHWorkload 				 health.KHWorkload
 }
 
 func init() {
@@ -151,6 +152,7 @@ func NewCheck(client *kubernetes.Clientset, checkConfig *khcheckcrd.Kuberhealthy
 		OriginalPodSpec:          checkConfig.Spec.PodSpec,
 		PodSpec:                  checkConfig.Spec.PodSpec,
 		KubeClient:               client,
+		KHWorkload:				  health.KHCheck,
 	}
 }
 
@@ -173,6 +175,7 @@ func NewJob(client *kubernetes.Clientset, jobConfig *khjobcrd.KuberhealthyJob, k
 		OriginalPodSpec:          jobConfig.Spec.PodSpec,
 		PodSpec:                  jobConfig.Spec.PodSpec,
 		KubeClient:               client,
+		KHWorkload:				  health.KHJob,
 	}
 }
 
@@ -368,7 +371,7 @@ func (ext *Checker) setUUID(uuid string) error {
 	// if the check was not found, we create a fresh one and start there
 	if err != nil && (k8sErrors.IsNotFound(err) || strings.Contains(err.Error(), "not found")) {
 		ext.log("khstate did not exist, so a default object will be created")
-		details := health.NewWorkloadDetails(checkState.Spec.GetKHWorkload())
+		details := health.NewWorkloadDetails(ext.KHWorkload)
 		details.Namespace = ext.CheckNamespace()
 		details.AuthoritativePod = ext.hostname
 		details.OK = true
